Remove BitPoints Integration and "squirrel crap"
The bitpoints integation is causing the HTTPS status for JIRA to show as
non secure. It doesn't appear to be used anymore so remove rather than
fix.

Also remove another commented out piece of code labeled as "squirrel
crap" as Flying Squirrels have not been a team for a while.